Add files via upload
Artemis will now output the command to run a Base64 Encoded bash reverse shell, in addition to the basic payload. This is useful if you need to URL-Encode a bash payload, & it can bypass some rudimentary Web App filters.

I also added a generic "Tiny PHP Payload " which can be used as a backdoor for Web Apps running on PHP servers.